Added Keychain functionality
The client side must authenticate with a password and a username to get the newest orders from the server. The password shall only be needed to be entered once. To save the password securly on the device the Keychain API must be used which stores the password encrypted on the device.