Now can log-in in Deliver App
The client app on the iOS device can now communicate with the server to check if a user account exists. Those user accounts will and can only be created on the server. The data must later be parsed with each request. Only if the username and password are correct the request is accepted. The username is stored as a CoreData entry. The password is stored securly on the device using keychain.