Store Order in Order App
App now stores an order to the devices storage after an successful request. The stored order consists of the data sent to the server and isn't validated with the server again after the successful order because this isn't needed unless some sent data is changed again on the server which should not happen by any means.