Added custom SQLiteDecimal TypeDecorator.
Added a SQLiteDecimal TypeDecorator because the pysqlite dialect doesn't support Decimal column types. This TypeDecorator converts floats to integers when saving these to a database and converts integers to floats when reading them from the database. The function does this like illustrated in following example: Given a SQLiteDecorator with the scale of 2 decimal places it would convert float 3.21 to 321 and from 321 back to 3.21 or convert 9.482 to 948 and from 948 back to 9.48.