Fixed order table bug.
Previously the name for the table which stores order information was order, which intercepted with the ORDER SQL keyword. This table is now called order_details.
ORM ForeignKeys are now mapped by taking the attribute of another ORM object, not anymore by parsing a string which indicates the foreign key relationship.