Now need to provide UACID when acquiring refresh token.
When a client tries to acquire a refresh token a UACID now has to be provided in the request headers. UACID stands for Unique Application Context ID. It is a randomly on-device generated identifier required to map refresh tokens to specific devices. The UACID is rotated each time the app is reinstalled and each time the user logs into the app (also after logging out). Thus it doesn't risk the privacy of the user as it can't be used for tracking or similar.