Background thread db storing now works for any data, not just orders.
A background thread listens for new items to be stored to the database. Therefor StoreOperation instances can be added to a queue. A StoreOperation consists out of a function which will be called from the background thread, positional arguments and keyword arguments. Previously this background thread was only intended to store orders.