Refresh token now gets stored in background thread.
When a new refresh token is issued it previously got written to the database before sending the response to the client. Now when a new refresh token is issued a store operation is created which will be executed in the background to store the refresh token to the database. Before sending the response the database now only has to be accessed for read operations (check if backend is allowed to issue new refresh token for specific user and UACID). The cost intensive write operations now don't anymore block the response of the request. Note: To return the refresh token it is generated before responding (but written in the background after responding).