Added "auth login" and "auth refresh" token info checks.
When a client performs a auth login or auth refresh the token information which gets generated is now checked. Those steps for example include checking if the maximum amount of refresh tokens were issued (auth login) or checking if an access token is already old enough to be replaced (auth refresh).