Added refresh token description.
Added a new table which stores extra information about a refresh token. Until now this only includes a device description. This is for example useful to see for which device a refresh token was initially created. This information is stored in this extra table as refresh tokens are invalidated and created very frequently. Hereby information which stays the same during the whole refresh token invalidation-creation process can be stored in this new table.