Added error code comments.
Added comments to the code where error codes are returned to the client to explain to the one reading why the request couldn't be full filled correctly. Detailed error descriptions isn't returned to the client as the backend is no public interface and thus information about errors would only help attackers.