Added new refresh token database design.
In a normal request-flow refresh tokens are newly issued each time access tokens are refreshed. All refresh tokens for one login are therefor chained together. For example: User logs in and gets a refresh token. -> Client updates access tokens with refresh token from login. He gets a new access token and a new refresh token. Hereby the refresh token is marked in the database as a successor of the first refresh token issued at login. This results in refresh token chains (which with this commit the database design now allows).