Make order requests now run through more detailed checks before responding.
Before responding new orders are now checked if their ordered items are valid. If those checks would be performed in the background like before it could be that a 200 response would be returned without knowing if the ordered items are valid. This could lead the client to interpreting the request/order as valid when it's clearly not.