Multiple improvments.
Removed the successful_response function as it didn't really has any purpose other than setting a "status": "successful" key on the response if it was a dictionary. As this state can be inferred from the HTTP response code the function isn't needed anymore.
Also made some improvements to the database design by setting a ON DELETE CASCADE key for ordered items if their associated order gets deleted.